refactor(workspace)!: split markdown helpers into their own subpath
Add @epicenter/workspace/markdown as a dedicated subpath for the
markdown file IO primitives. The materializer subpath
@epicenter/workspace/document/materializer/markdown now publishes one
coherent product (the materializer + its filename slot factory); the
new subpath publishes another (parse/emit ---fenced YAML-frontmatter
files plus the shared slug-filename convention).

  @epicenter/workspace/markdown
    parseMarkdownFile
    assembleMarkdown
    toSlugFilename
    SerializeResult

  @epicenter/workspace/document/materializer/markdown  (unchanged shape)
    attachMarkdownMaterializer
    slugFilename
    MaterializerPushError, PushEvent, PushResult

This replaces the previous commit a5089051b approach (delete the
helpers from the workspace, duplicate into the playground), which hid
the cohesion smell rather than fixing it. The earlier audit framed the
goal as "shrink the public surface"; the correct framing per
cohesive-clean-breaks is "split two products into two APIs." The
playground keeps prepareMarkdownFiles locally because it encodes one
vault layout (flat directory, `id` field, abort on collision) and is
not a workspace primitive. toIdFilename is dropped: it was a one-line
inline (\`${id}.md\`) with one internal caller.

BREAKING CHANGE: parseMarkdownFile, assembleMarkdown, and toSlugFilename
now import from @epicenter/workspace/markdown instead of from
@epicenter/workspace/document/materializer/markdown. toIdFilename is
removed; inline as \`${id}.md\` if needed. prepareMarkdownFiles stays
unexported from the workspace; copy the playground implementation if
you need the same vault recipe.
