+++ packages/workspace/package.json
@@ -27,6 +27,7 @@
 	"types": "./src/index.ts",
 	"exports": {
 		".": "./src/index.ts",
+		"./markdown": "./src/markdown/index.ts",
 		"./document/materializer/markdown": "./src/document/materializer/markdown/index.ts",
 		"./daemon": "./src/daemon/index.ts",
 		"./ai": "./src/ai/index.ts",
