+++ packages/workspace/src/document/materializer/markdown/serializers.ts
@@ -1,28 +1,5 @@
-import slugify from '@sindresorhus/slugify';
-import filenamify from 'filenamify';
+import { toSlugFilename } from '../../../markdown/to-slug-filename.js';
 import type { BaseRow } from '../../attach-table.js';
-
-/** Max slug length before the ID suffix. */
-const MAX_SLUG_LENGTH = 50;
-
-/** Build an ID-only filename: `{id}.md`. */
-function toIdFilename(id: string): string {
-	return `${id}.md`;
-}
-
-/**
- * Build a `{slugified-title}-{id}.md` filename, falling back to `{id}.md`
- * when the title is empty.
- */
-function toSlugFilename(title: string | undefined | null, id: string): string {
-	if (!title || title.trim().length === 0) {
-		return toIdFilename(id);
-	}
-
-	const slug = slugify(title).slice(0, MAX_SLUG_LENGTH);
-	const raw = slug ? `${slug}-${id}.md` : toIdFilename(id);
-	return filenamify(raw, { replacement: '-' });
-}
 
 /**
  * Build a `filename` slot that produces `{slug}-{id}.md` using a row field
